// sdk/src/generated/models/PaginatedUserGroups.ts
/**
 * Stack Overflow for Teams API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { SortOrder } from '../models/SortOrder';
import { UserGroupResponseModel } from '../models/UserGroupResponseModel';
import { UserGroupsSortParameter } from '../models/UserGroupsSortParameter';
import { HttpFile } from '../http/http';

export class PaginatedUserGroups {
    /**
    * Total count. The total number of records found.
    */
    'totalCount'?: number;
    /**
    * Page size. The number of records to retrieve per page.
    */
    'pageSize'?: number;
    /**
    * Page. The number of the current page.
    */
    'page'?: number;
    /**
    * Total pages. The total number of pages available.
    */
    'totalPages'?: number;
    'sort'?: UserGroupsSortParameter;
    'order'?: SortOrder;
    'items'?: Array<UserGroupResponseModel>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "totalCount",
            "baseName": "totalCount",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "pageSize",
            "baseName": "pageSize",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "page",
            "baseName": "page",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "totalPages",
            "baseName": "totalPages",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "sort",
            "baseName": "sort",
            "type": "UserGroupsSortParameter",
            "format": ""
        },
        {
            "name": "order",
            "baseName": "order",
            "type": "SortOrder",
            "format": ""
        },
        {
            "name": "items",
            "baseName": "items",
            "type": "Array<UserGroupResponseModel>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return PaginatedUserGroups.attributeTypeMap;
    }

    public constructor() {
    }
}