// sdk/src/generated/models/UserGroupResponseModel.ts
/**
 * Stack Overflow for Teams API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { UserSummaryResponseModel } from '../models/UserSummaryResponseModel';
import { HttpFile } from '../http/http';

export class UserGroupResponseModel {
    /**
    * ID. The user group\'s unique identifier.
    */
    'id'?: number;
    /**
    * Name. The user group\'s name.
    */
    'name'?: string;
    /**
    * Description. The user group\'s description.
    */
    'description'?: string;
    'users'?: Array<UserSummaryResponseModel>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string",
            "format": ""
        },
        {
            "name": "users",
            "baseName": "users",
            "type": "Array<UserSummaryResponseModel>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return UserGroupResponseModel.attributeTypeMap;
    }

    public constructor() {
    }
}